Add shared next-config library for Next.js applications
Extract common Next.js configuration into a reusable library to standardize
settings across all Next.js applications in the workspace. This includes
ESLint/TypeScript ignore settings and GitHub Actions optimizations.
