+++ .nx/version-plans/add-next-config-library.md
@@ -1,0 +1,5 @@
+---
+__default__: minor
+---
+
+Add shared next-config library to standardize Next.js configurations across all applications
