+++ apps/example-next-js/next.config.js
@@ -1,0 +1,3 @@
+import nextConfig from "next-config";
+
+export default nextConfig;
