+++ apps/example-vercel-ts/package.json
@@ -7,6 +7,7 @@
     "@modelcontextprotocol/sdk": "^1.16.0",
     "@modelfetch/vercel": "workspace:^",
     "next": "^15.4.1",
+    "next-config": "workspace:^",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "zod": "^3.25.76"
