+++ apps/modelfetch-website/next.config.js
@@ -1,20 +1,4 @@
 import { createMDX } from "fumadocs-mdx/next";
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  eslint: { ignoreDuringBuilds: true },
-  typescript: { ignoreBuildErrors: true },
-};
-
-if (process.env.GITHUB_ACTIONS === "true") {
-  nextConfig.webpack = (config) => {
-    if (config.cache) config.cache = Object.freeze({ type: "memory" });
-    return config;
-  };
-  nextConfig.experimental = {
-    webpackMemoryOptimizations: true,
-    webpackBuildWorker: true,
-  };
-}
+import nextConfig from "next-config";
 
 export default createMDX()(nextConfig);
