+++ apps/modelfetch-website/package.json
@@ -10,6 +10,7 @@
     "fumadocs-ui": "^15.6.4",
     "lucide-react": "^0.525.0",
     "next": "^15.4.1",
+    "next-config": "workspace:^",
     "next-themes": "^0.4.6",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
