+++ libs/next-config/README.md
@@ -1,0 +1,12 @@
+# `next-config`
+
+> [!NOTE]
+> This is an internal package.
+
+## Usage
+
+```javascript
+import nextConfig from "next-config";
+
+export default nextConfig;
+```
