+++ libs/next-config/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "next-config",
+  "version": "0.13.7",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.js"
+}
